requirements.txt
aiofiles==24.1.0
aiogram==2.25.1
aiohappyeyeballs==2.6.1
aiohttp==3.8.6
aiosignal==1.3.2
annotated-types==0.7.0
async-timeout==4.0.3
attrs==25.3.0
Babel==2.9.1
certifi==2025.1.31
charset-normalizer==3.4.1
frozenlist==1.5.0
idna==3.10
magic-filter==1.0.12
multidict==6.3.2
propcache==0.3.1
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.1.0
pytz==2025.2
requests==2.32.3
typing_extensions==4.13.1
urllib3==2.3.0
yarl==1.18.3